Update blazy.js
Blazy currently uses “unshift” which is not supported by IE8 and earlier according to http://www.w3schools.com/jsref/jsref_unshift.asp. Since there is an IE7 workaround for “querySelectorAll”, those versions were supposed to be supported. It can be replaced by “push” which is supported. The only difference is the order of lazy loading.